+++ libraries/botbuilder-middleware-text-analytics/src/schema.ts
@@ -1,0 +1,10 @@
+import { ServiceClientOptions } from '@azure/ms-rest-js';
+
+/**
+ * @module botbuildercommunity/middleware-text-analytics
+ */
+
+export interface TextAnalysisOptions extends ServiceClientOptions {
+    serviceKey: string;
+    endpoint: string;
+}